requirements.txt
yara-python
flask
psutil
pywin32
scikit-learn